GSK-4 Java migration: solve pr comments
