+++ giskard-server/src/main/java/ai/giskard/web/rest/controllers/ModelsController.java
@@ -1,6 +1,5 @@
 package ai.giskard.web.rest.controllers;
 
-import ai.giskard.domain.ml.ProjectModel;
 import ai.giskard.repository.ProjectRepository;
 import ai.giskard.repository.UserRepository;
 import ai.giskard.repository.ml.ModelRepository;
