+++ giskard-server/src/main/java/ai/giskard/web/rest/errors/Entity.java
@@ -1,0 +1,18 @@
+package ai.giskard.web.rest.errors;
+
+import lombok.Getter;
+
+public enum Entity {
+    PROJECT("Project"),
+    PROJECT_MODEL("Model"),
+    TEST_SUITE("Test suite"),
+    TEST("Test"),
+    USER("User");
+
+    @Getter
+    private final String name;
+
+    Entity(String name) {
+        this.name = name;
+    }
+}
